feat" create subscription table upon user sing up
