+++ requirements.txt
@@ -6,7 +6,9 @@ botocore==1.34.11
 certifi==2023.11.17
 cffi==1.16.0
 charset-normalizer==3.3.2
+contourpy==1.2.0
 cryptography==41.0.7
+cycler==0.12.1
 defusedxml==0.7.1
 dj-rest-auth==5.0.2
 Django==5.0
@@ -15,10 +17,14 @@ django-redis==5.4.0
 django-storages==1.14.2
 djangorestframework==3.14.0
 ecdsa==0.18.0
+fonttools==4.47.0
 idna==3.6
 jmespath==1.0.1
+kiwisolver==1.4.5
+matplotlib==3.8.2
 numpy==1.26.2
 oauthlib==3.2.2
+packaging==23.2
 pandas==2.1.4
 paramiko==3.4.0
 Pillow==10.1.0
@@ -28,6 +34,8 @@ pycparser==2.21
 pygame==2.5.2
 PyJWT==1.7.1
 PyNaCl==1.5.0
+pyparsing==3.1.1
+pysftp==0.2.9
 python-dateutil==2.8.2
 python-dotenv==1.0.0
 python-jose==3.3.0
